Safari-609.1.11.1/LayoutTests/storage/websql/private-browsing-noread-nowrite-expected.txt

CONSOLE MESSAGE: line 92: Web SQL is deprecated. Please use IndexedDB instead.
This test makes sure that attempts to change the database during private browsing fail.
Setup statement 1 completed successfully
Setup statement 2 completed successfully
Private browsing statement 1 completed with an error
could not prepare statement (1 not authorized)
Private browsing statement 2 completed with an error
could not prepare statement (23 not authorized)
Private browsing statement 3 completed with an error
could not prepare statement (23 not authorized)
Private browsing statement 4 completed with an error
could not prepare statement (23 not authorized)
Private browsing statement 5 completed with an error
could not prepare statement (23 access to PrivateTest1.randomData is prohibited)
Test ended
